install scripts: avoid to run twice

+++ toucan_connectors/install_scripts/azure_mssql.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 set -e
 
+if [[ -e ~/azure_mssql-installed ]]; then
+    echo "Azure MSSQL connector dependencies are already installed."
+    exit
+fi
+
 curl https://packages.microsoft.com/keys/microsoft.asc | apt-key add -
 curl https://packages.microsoft.com/config/ubuntu/14.04/prod.list | tee /etc/apt/sources.list.d/mssql-release.list
 apt-get update
 ACCEPT_EULA=Y apt-get -y install msodbcsql17 unixodbc-dev
+
+touch ~/azure_mssql-installed
