+++ toucan_connectors/install_scripts/oracle.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -e
 
+if [[ -e ~/oracle-installed ]]; then
+    echo "Oracle connector dependencies are already installed"
+    exit
+fi
+
 apt-get install -fyq libaio1 curl
 mkdir -p /opt/oracle
 curl -s 'https://raw.githubusercontent.com/circulosmeos/gdown.pl/master/gdown.pl' -o /tmp/gdown.pl
@@ -10,3 +15,5 @@ unzip /tmp/oracle_client_lib.zip -d /opt/oracle
 sh -c "echo /opt/oracle/instantclient_12_2 > /etc/ld.so.conf.d/oracle-instantclient.conf"
 ldconfig
 rm -rf /tmp/gdown.pl /tmp/oracle_client_lib.zip
+
+touch ~/oracle-installed
